Fix dependency on neon-image in promote-images-dev (#10437)
## Problem
72d544568eb3d2ea3be0b4afb1ba24c804e5e6fa failed CI on main because a job
ran to soon. This was caused by
019eabe6f5ad16d2ee0405808c21ee7bbcc57f1b. While `promote-images-dev`
does not inherently need `neon-image`, a few jobs depending on
`promote-images-dev` do need it, and previously had it when it was
`promote-images`, which depended on `test-images`, which in turn
depended on `neon-image`.

## Summary of changes
To ensure jobs depending `docker.io/neondatabase/neon` images get them,
`promote-images-dev` gets the dependency to `neon-image` back which it
previously had transitively through `test-images`.